refactor(web): move add linked appeal routes into a subrouter
