+++ appeals/web/src/server/appeals/appeal-details/linked-appeals/add/add.middleware.js
@@ -1,0 +1,5 @@
+/** @type {import('express').RequestHandler} */
+export function initialiseLinkedAppealsSession(request, _, next) {
+	request.session.linkableAppeal ??= {};
+	next();
+}
