+++ appeals/web/src/server/appeals/appeal-details/linked-appeals/linked-appeals.router.js
@@ -3,30 +3,13 @@ import { asyncHandler } from '@pins/express';
 import * as controller from './linked-appeals.controller.js';
 import * as validators from './linked-appeals.validators.js';
 import { initialiseLinkedAppealsSession } from './linked-appeals.middleware.js';
+import addRouter from './add/add.router.js';
 
 const router = createRouter({ mergeParams: true });
 
 router.use(initialiseLinkedAppealsSession);
 
-router
-	.route('/add')
-	.get(asyncHandler(controller.renderAddLinkedAppealReference))
-	.post(validators.validateAddLinkedAppealReference, asyncHandler(controller.postAddLinkedAppeal));
-
-router
-	.route('/add/already-linked')
-	.get(asyncHandler(controller.renderAlreadyLinked))
-	.post(asyncHandler(controller.postAlreadyLinked));
-
-router
-	.route('/add/lead-appeal')
-	.get(asyncHandler(controller.renderLeadAppeal))
-	.post(validators.validateChangeLeadAppeal, asyncHandler(controller.postLeadAppeal));
-
-router
-	.route('/add/check-and-confirm')
-	.get(asyncHandler(controller.renderAddLinkedAppealCheckAndConfirm))
-	.post(asyncHandler(controller.postAddLinkedAppealCheckAndConfirm));
+router.use('/add', addRouter);
 
 router.route('/manage').get(asyncHandler(controller.renderManageLinkedAppeals));
